% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_data_HYDRO}
\alias{create_data_HYDRO}
\title{Create data}
\usage{
create_data_HYDRO(
  computer_data_path,
  filedir = "",
  filename = "all",
  variable_to_load = c("Qm3s", "Qls", "Qmmj"),
  val_to_keep = NULL,
  format = "HYDRO2",
  verbose = TRUE
)
}
\arguments{
\item{computer_data_path}{Path to the data.}

\item{filedir}{Directory of Banque HYDRO data you want to use in
ASHE\\computer_data_path\\ to get station codes. If "" is
use, data will be search in ASHE\\computer_data_path\\.}

\item{filename}{String or vector of string of all filenames from
which data will be created. If set to 'all', all the file in
'fildir' will be use.}

\item{verbose}{Boolean to indicate if more processing info are
printed (default : TRUE).}
}
\value{
A tibble containing data about selected stations.
}
\description{
Creation of data of stations.
}
