% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_data_HYDRO3}
\alias{create_data_HYDRO3}
\title{create_data_HYDRO3}
\usage{
create_data_HYDRO3(
  paths,
  variable_to_load = c("Qm3s", "Qls", "Qmmj"),
  value_to_keep = NULL,
  verbose = FALSE
)
}
\description{
create_data_HYDRO3
}
