% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_meta_HYDRO2}
\alias{create_meta_HYDRO2}
\title{Create metadata}
\usage{
create_meta_HYDRO2(
  computer_data_path,
  filedir,
  filename,
  hydrological_region_level = 1,
  verbose = TRUE
)
}
\arguments{
\item{computer_data_path}{Path to the data.}

\item{filedir}{Directory of Banque HYDRO data you want to use in
ASHE\\computer_data_path\\ to get station codes. If "" is
use, data will be search in ASHE\\computer_data_path\\.}

\item{filename}{String or vector of string of all filenames from
which metadata will be created. If set to 'all', all the file in
'fildir' will be use.}

\item{verbose}{Boolean to indicate if more processing info are
printed (default : TRUE).}
}
\value{
A tibble containing metadata about selected stations.
}
\description{
Creation of metadata of stations.
}
