% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_tools.R
\name{get_alphabet_in_px}
\alias{get_alphabet_in_px}
\title{get_alphabet_in_px}
\usage{
get_alphabet_in_px(
  alphabet = c(letters, LETTERS, c("é", "è", "à", "ç", "É", "È", "À", "Ç"),
    c("1", "2", "3", "4", "5", "6", "7", "8", "9", "0"), c("-", "_", ".", ",", "*", "'",
    "\%", "(", ")", "[", "]", "{", "}", "!", "?", "+", "=", "@", "|", "#", "&")),
  size = 50,
  font = "sans",
  style = "normal",
  isNorm = TRUE,
  out_dir = "letter",
  save = FALSE
)
}
\description{
get_alphabet_in_px
}
