% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_color.R
\name{theme_IPCC}
\alias{theme_IPCC}
\title{Ggplot2 theme ash}
\usage{
theme_IPCC(
  isBack = TRUE,
  isGridX = FALSE,
  isGridY = TRUE,
  isTitle = FALSE,
  dTitle = 0,
  isLabelX = FALSE,
  isLabelY = FALSE
)
}
\description{
Ggplot2 theme ash
}
