% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_tibble}
\alias{read_tibble}
\title{read_tibble}
\usage{
read_tibble(filepath = NULL, filedir = "./", filename = "data.txt")
}
\arguments{
\item{filepath}{\emph{character, default=NULL} Path to the file to read.}

\item{filedir}{\emph{character, default="./"} Directory to the file to read.}

\item{filename}{\emph{character, default="data.txt"} Name of the file to read.}
}
\description{
Reads a file previously writed with \code{\link[=write_tibble]{write_tibble()}} and return a \code{\link[dplyr:reexports]{dplyr::tibble()}} or a list of \code{\link[dplyr:reexports]{dplyr::tibble()}}.
}
\details{
This set of \code{\link[=read_tibble]{read_tibble()}} and \code{\link[=write_tibble]{write_tibble()}} functions support \code{.Rdata}, \code{.txt} and \code{.fst} format.
}
\examples{
```
# load data
data("iris")
# convert data
iris = dplyr::tibble(iris)
iris$Species = as.character(iris$Species)

# writes a tibble
write_tibble(iris, filedir="./", filename="iris.txt")
# or a list of tibble
iris_setosa = iris[iris$Species == "setosa",]
iris_virginica = iris[iris$Species == "virginica",]
iris_selection = list(setosa=iris_setosa, virginica=iris_virginica)
write_tibble(iris_selection, filedir="./", filename="iris_selection.txt")

# read with a path
read_tibble(filepath="iris.txt")
# or with a directory and a filename
read_tibble(filedir="./", filename="iris_selection.txt")
```
}
