% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_tibble}
\alias{write_tibble}
\title{write_tibble}
\usage{
write_tibble(tbl, filedir = "./", filename = "data.txt")
}
\arguments{
\item{tbl}{\code{\link[dplyr:reexports]{dplyr::tibble()}} to write.}

\item{filedir}{\emph{tibble, default="./"}   Directory to write the file.}

\item{filename}{\emph{character, default="data.txt"}   Name of the file to write.}
}
\description{
Writes a \code{\link[dplyr:reexports]{dplyr::tibble()}} or a list of \code{\link[dplyr:reexports]{dplyr::tibble()}}.
}
\details{
This set of \code{\link[=read_tibble]{read_tibble()}} and \code{\link[=write_tibble]{write_tibble()}} functions support \code{.Rdata}, \code{.txt} and \code{.fst} format.
}
\examples{
# load data
data("iris")
# convert data
iris = dplyr::tibble(iris)
iris$Species = as.character(iris$Species)

# writes a tibble
write_tibble(iris, filedir="./", filename="iris.txt")
# or a list of tibble
iris_setosa = iris[iris$Species == "setosa",]
iris_virginica = iris[iris$Species == "virginica",]
iris_selection = list(setosa=iris_setosa, virginica=iris_virginica)
write_tibble(iris_selection, filedir="./", filename="iris_selection.txt")
}
